+++ app/src/main/java/com/alican/mvvm_starter/data/model/MovieResponseModel.kt
@@ -13,6 +13,6 @@ data class MovieResponseModel(
     val release_date: String,
     val title: String,
     val video: Boolean,
-    val vote_average: Float,
+    val vote_average: Double,
     val vote_count: Int
 )
